Finished the PR comments made on this issue. Also refactored the names of the functions from updateTrack to just update. Fixed the test_userRsvPForEvent that was failing whenever we ran our test cases.
